refactor: rename data-table to tasks-table
